Move FlattenedStructureContainer to own module
also implements HasStructure for it
