+++ pyiron_contrib/atomistics/atomicrex/structure_list.py
@@ -6,6 +6,7 @@ from ase import Atoms as ASEAtoms
 from pyiron_base import DataContainer
 from pyiron_atomistics import pyiron_to_ase, ase_to_pyiron, Atoms
 
+from pyiron_contrib.atomistics.job.structurecontainer import StructureContainer
 from pyiron_contrib.atomistics.atomicrex.fit_properties import ARFitPropertyList, ARFitProperty
 from pyiron_contrib.atomistics.atomicrex.utility_functions import write_pretty_xml
 from pyiron_contrib.atomistics.atomicrex.fit_properties import FlattenedARProperty, FlattenedARVectorProperty
@@ -17,160 +18,6 @@ from pyiron_contrib.atomistics.atomicrex.fit_properties import FlattenedARProper
 #        self.cell = cell
 #        self.symbols = symbols
 #        self.positions = positions
-
-
-class FlattenedStructureContainer:
-    """
-    Class that can write and read lots of structures from and to hdf quickly.
-
-    This is done by storing positions, cells, etc. into large arrays instead of writing every structure into a new
-    group.  Structures are stored together with an identifier that should be unique.  The class can be initialized with
-    the number of structures and the total number of atoms in all structures, but re-allocates memory as necessary when
-    more (or larger) structures are added than initially anticipated.
-    """
-
-    def __init__(self, num_structures=1, num_atoms=1):
-        """
-        Create new structure container.
-
-        Args:
-            num_structures (int): pre-allocation for per structure arrays
-            num_atoms (int): pre-allocation for per atoms arrays
-        """
-        # tracks allocated versed as yet used number of structures/atoms
-        self._num_structures_alloc = self.num_structures = num_structures
-        self._num_atoms_alloc = self.num_atoms = num_atoms
-        # store the starting index for properties with unknown length
-        self.current_atom_index = 0
-        # store the index for properties of known size, stored at the same index as the structure
-        self.current_structure_index = 0
-        # Also store indices of structure recently added
-        self.prev_structure_index = 0
-        self.prev_atom_index = 0
-
-        self._init_arrays()
-
-
-    def _init_arrays(self):
-        self.symbols = np.full(self._num_atoms_alloc, "XX", dtype=np.dtype("U2"))# 2 character unicode array for chemical symbols
-        self.positions = np.empty((self._num_atoms_alloc, 3))
-        self.cells = np.empty((self._num_structures_alloc, 3, 3))
-        self.start_indices = np.empty(self._num_structures_alloc, dtype=np.int32)
-        self.len_current_struct = np.empty(self._num_structures_alloc, dtype=np.int32)
-        self.identifiers = np.empty(self._num_structures_alloc, dtype=np.dtype("U20"))
-
-    def _resize_atoms(self, new):
-        self._num_atoms_alloc = new
-        try:
-            self.symbols.resize(new)
-            self.positions.resize( (new, 3) )
-        except ValueError:
-            self.symbols = np.resize(self.symbols, new)
-            self.positions = np.resize(self.positions, (new, 3) )
-
-    def _resize_structures(self, new):
-        self._num_structures_alloc = new
-        try:
-            self.cells.resize( (new, 3, 3) )
-            self.start_indices.resize(new)
-            self.len_current_struct.resize(new)
-            self.identifiers.resize(new)
-        except ValueError:
-            self.cells = np.resize(self.cells, (new, 3, 3) )
-            self.start_indices = np.resize(self.start_indices, new)
-            self.len_current_struct = np.resize(self.len_current_struct, new)
-            self.identifiers = np.resize(self.identifiers, new)
-
-    def add_structure(self, structure, identifier):
-        n = len(structure)
-        new_atoms = self.current_atom_index + n
-        if new_atoms > self._num_atoms_alloc:
-            self._resize_atoms(max(new_atoms, self._num_atoms_alloc * 2))
-        if self.current_structure_index + 1 > self._num_structures_alloc:
-            self._resize_structures(self._num_structures_alloc * 2)
-
-        if new_atoms > self.num_atoms:
-            self.num_atoms = new_atoms
-        if self.current_structure_index + 1 > self.num_structures:
-            self.num_structures += 1
-
-        # len of structure to index into the initialized arrays
-        i = self.current_atom_index + n
-
-        self.symbols[self.current_atom_index:i] = np.array(structure.symbols)
-        self.positions[self.current_atom_index:i] = structure.positions
-
-        self.len_current_struct[self.current_structure_index] = n
-        self.cells[self.current_structure_index] = structure.cell.array
-        self.start_indices[self.current_structure_index] = self.current_atom_index
-        self.identifiers[self.current_structure_index] = identifier
-
-        self.prev_structure_index = self.current_structure_index
-        self.prev_atom_index = self.current_atom_index
-
-        # Set new current_atom_index and increase current_structure_index
-        self.current_structure_index += 1
-        self.current_atom_index = i
-        #return last_structure_index, last_atom_index
-
-
-    def to_hdf(self, hdf, group_name="flattened_structures"):
-        # truncate arrays to necessary size before writing
-        self._resize_atoms(self.num_atoms)
-        self._resize_structures(self.num_structures)
-
-        with hdf.open(group_name) as hdf_s_lst:
-            hdf_s_lst["symbols"] = self.symbols.astype(np.dtype("S2"))
-            hdf_s_lst["positions"] = self.positions
-            hdf_s_lst["cells"] = self.cells
-            hdf_s_lst["start_indices"] = self.start_indices
-            hdf_s_lst["identifiers"] = self.identifiers.astype(np.dtype("S20"))
-            hdf_s_lst["num_atoms"] =  self._num_atoms_alloc
-            hdf_s_lst["num_structures"] = self._num_structures_alloc
-            hdf_s_lst["len_current_struct"] = self.len_current_struct
-
-
-    def from_hdf(self, hdf, group_name="flattened_structures"):
-        with hdf.open(group_name) as hdf_s_lst:
-            self._num_structures_alloc = hdf_s_lst["num_structures"]
-            self._num_atoms_alloc = hdf_s_lst["num_atoms"]
-
-            self._init_arrays()
-
-            self.symbols = hdf_s_lst["symbols"].astype(np.dtype("U2"))
-            self.positions = hdf_s_lst["positions"]
-            self.cells = hdf_s_lst["cells"]
-            self.start_indices = hdf_s_lst["start_indices"]
-            self.identifiers = hdf_s_lst["identifiers"].astype(np.dtype("U20"))
-            self.len_current_struct = hdf_s_lst["len_current_struct"]
-
-    def get_structure(self, index_or_identifier=-1):
-        """
-        Create pyiron Atoms object from saved arrays.
-
-        Args:
-            index_or_identifier (int, str): if str then must one of the identifiers given to :method:`.add_structure`
-                                            otherwise must be and integer in range [-N, N) where N the number of
-                                            structures stored in the container
-
-        Returns:
-            :class:`pyiron_atomistics.atomistics.structure.atoms.Atoms`: requested structure
-        """
-        if isinstance(index_or_identifier, str):
-            try:
-                index = np.where(self.identifiers == index_or_identifier)[0][0]
-            except IndexError:
-                raise IndexError(f"No structure named {index_or_identifier} found in structure container.") from None
-        else:
-            index = index_or_identifier
-        if index < 0:
-            index += self.num_structures
-        if not (0 <= index < self.num_structures):
-            raise IndexError(f"Index {index} not in in range [-{self.num_structures}, {self.num_structures})")
-        I = self.start_indices[index]
-        E = I + self.len_current_struct[index]
-        return Atoms(symbols=self.symbols[I:E], cell=self.cells[index], positions=self.positions[I:E])
-
 
 
 class ARStructureContainer:
